Parse Kalshi's new positions response shape
/portfolio/positions rows now carry 'position_fp' (string float, can be
fractional) and '_dollars' string fields instead of the old integer
'position' / 'market_exposure' / 'realized_pnl'. Our parser was looking
only for the old keys, so every row came back as position=0 and got
filtered, producing an empty /status even when the account held seven
open positions.

- _parse_position() reads 'position' first, then 'position_fp'; returns
  a signed float (YES positive, NO negative) so fractional markets work.
- _cents_from_row() reads '<name>' as int cents first, then falls back
  to '<name>_dollars' as a string-dollar value rounded to cents.
- Position.position is now float; formatter renders it with 2 decimals
  only when actually fractional.

Backward-compatible with the old integer response shape.

+++ src/kalshi_tradr/kalshi/types.py
@@ -50,7 +50,7 @@ class OrderResult:
 @dataclass(slots=True)
 class Position:
     ticker: str
-    position: int
+    position: float  # signed contracts; positive = YES, negative = NO. Fractional in some markets.
     market_exposure: int  # cents
     realized_pnl: int  # cents
     raw: dict[str, Any] = field(repr=False, default_factory=dict)
